requirements.txt
pyserial
PyBoard
esptool